fix(trips): the preview takes a fractional distance, so a new row keeps its ODO
The preview command declared `distance_km` as an i32. A km like 100.5 therefore
failed `parse_args`, the request 400'd, and the editor lost its answer.

On this branch that costs a legal value. The row editor now reads its odometer
from the preview, so a null response leaves a new row with no `odometerAnchor`,
`handleSave`'s clamp cannot fire, and `formData.odometer` -- null on a new row --
saves as 0. Before the branch the local `previousOdometer + km` covered it, and
only the live preview was lost.

Widen `distance_km` to f64, the type `Trip.distance_km` carries everywhere else.
That removes the failure class rather than the symptom. `current_trip_km` on the
magic-fill arm was already f64, so the i32 was the single outlier and nothing
else moved. `api.ts` already typed the argument `number`.

Two comments went stale and are corrected here:
  - the km-match guard in `applyPreviewOdometer` claimed the command truncates a
    fractional km to i32. It never truncated -- it rejected the request. The
    guard now does two things, not three, and it accepts a whole or a fractional
    km alike;
  - `TripGrid`'s display-order note still described a two-key rule, against the
    four-key `trip_order` the backend ships.

Tests:
  - `test_preview_answers_a_fractional_distance` asserts odometerStart 50050 and
    odometer 50150.5 for a 100.5 km preview. Against the i32 it did not compile;
  - an integration case types 100.5 km into a new row and reads the STORED
    odometer back over RPC, because the grid cell rounds to whole km. With the
    widening reverted it reports `Expected: 80200.5 / Received: 0`.

+++ src-tauri/core/src/server/dispatcher.rs
@@ -368,7 +368,10 @@ pub fn dispatch_sync(command: &str, args: Value, state: &ServerState) -> Result<
             struct Args {
                 vehicle_id: String,
                 year: i32,
-                distance_km: i32,
+                // f64, like `Trip.distance_km` everywhere else. An i32 here
+                // rejected a fractional km outright, so the preview never
+                // answered and a new row saved odometer 0.
+                distance_km: f64,
                 fuel_liters: Option<f64>,
                 full_tank: bool,
                 insert_at_trip_id: Option<String>,
